Replace new constructors with private functions
